+++ src/click/types.py
@@ -157,7 +157,14 @@ class Choice(ParamType):
         self.case_sensitive = case_sensitive
 
     def get_metavar(self, param):
-        return f"[{'|'.join(self.choices)}]"
+        choices_str = "|".join(self.choices)
+
+        # Use curly braces to indicate a required argument.
+        if param.required and param.param_type_name == "argument":
+            return f"{{{choices_str}}}"
+
+        # Use square braces to indicate an option or optional argument.
+        return f"[{choices_str}]"
 
     def get_missing_message(self, param):
         choice_str = ",\n\t".join(self.choices)