+++ CHANGES.rst
@@ -67,6 +67,9 @@ Unreleased
 -   The result object returned by the test runner's ``invoke()`` method
     has a ``return_value`` attribute with the value returned by the
     invoked command. :pr:`1312`
+-   Required arguments with the ``Choice`` type show the choices in
+    curly braces to indicate that one is required (``{a|b|c}``).
+    :issue:`1272`
 
 
 Version 7.1.2